Remove 1.3 and 1.4 data.
The new doc parser can't read the old data, and data shouldn't be checked into git anyway.
This removes everything from the dropdown except for 1.8, which is fine with me.   The
main dojotoolkit.org page can have a link for "old API doc" pointing to the old viewer,
the same way that it's supposed to have a link to "old reference doc".

Refs #8.